Crate coreutils_core

Re-exports

pub use bstr;

pub use libc;

pub use time;

Modules

backup
The Backup module handles creating backups using the methods specified in the GNU coreutils manual.

consts
CFG constants

env
Module for environments abstractions.

input
The Input module handles getting user input from the command line

mkfifo
Helper functions to deal with FIFO special file.

mktemp
Module with wrappers for libc mkstemp(3), mkdtemp(3).

os
Module for OS processes and types abstractions.

Structs

BStr
A wrapper for &[u8] that provides convenient string oriented trait impls.

BString
A wrapper for Vec<u8> that provides convenient string oriented trait impls.

Traits

ByteSlice
A trait that extends &[u8] with string oriented methods.

ByteVec
A trait that extends Vec<u8> with string oriented methods.

Functions

B
A short-hand constructor for building a &[u8].