Function coreutils_core::backup::create_existing_backup

pub fn create_existing_backup(
    file: &Path, 
    suffix: &str
) -> Result<PathBuf, Error>

Creates a backup in-keeping with previous backups. Pokes the directory to see whether there are any numbered or simple backups of the input file. If numbered backups are found, a numbered backup will be created. Else, a simple backup is created using the input suffix

Errors

If this function encounters any kind of I/O error, an error variant will be returned.