Module coreutils_core::os

Module for OS processes and types abstractions.

Modules

group
Module to deal more easily with UNIX groups.

load
Module related to system load.

login_name
Module to abstract libc function that get user login name.

passwd
Module to deal more easily with UNIX passwd.

process
Module with abstractions to safely deal with processes.

time
Module for time related abstractions more close to the OS.

tty
Module for TTY abstractions.

utmpx
Extended account database module.

utsname
Module for system information.

Functions

get_effective_gid
Get the current running process user effective group id.

get_effective_uid
Get the current running process user effective user id.

get_real_gid
Get the current running process user real group id.

get_real_uid
Get the current running process user real user id.

Type Definitions

Fields
Passwd field type.

Gid
Group ID type.

Pid
Process ID Type.

Susec
Field for TimeVal in microseconds.

Time
Passwd time type.

TimeVal
Time stamp type used on system structures.

Tm

Uid
User ID type.