Module coreutils_core::os::group

Module to deal more easily with UNIX groups.

Structs

Group
This struct holds information about a group of UNIX/UNIX-like systems.

Groups
A collection of Group.

Enums

Error
Enum that holds possible errors while creating Group type.

Type Definitions

Members
A iterator of group members.

Result